Add "backup" option to "put" methods to allow them to backup existing datasets when writing.
Note that the real implementation for this feature is in daf_butlerUtils's CameraMapper class.
The backup feature is *not* safe in parallel processing, as I don't think there's a way to protect it from race conditions at this level, but this shouldn't matter for the main use case (saving configs); see pipe.base.CmdLineTask.writeConfig for more information.