prevent opening files that are already open
tempfile.NamedTemporaryFile holds the file open, but we have been reopening
it (using the filename to get a file pointer); this can cause problems on
some systems.

Changed SafeFile to return a python file object, to avoid reopening the file.
Added SafeFilename to return a filename that doesn't have an open fd. This is
necessary to support FITS files, as cfitsio doesn't have an interface that
accepts a FILE* or fd.

Conflicts:
	python/lsst/daf/persistence/butler.py
	python/lsst/daf/persistence/eupsVersions.py

This is essentially a cherry-pick of HSC-1292's:
https://github.com/HyperSuprime-Cam/daf_persistence/commit/e526b5bf93111a4090c3ddb22b51f40c13129c80

Note that the
if storageName == 'FitsStorage':
persistence clause has been added to the put() function here.  It was
originally introduced in HSC on commit:
https://github.com/HyperSuprime-Cam/daf_persistence/commit/e37f6603b49b8ee5347fc4d2ca11500d25041276
which has not (yet?) been pulled over.