Implements butler datasetType aliases and related unit tests
After instantiating a butler, you can call registerDatasetTypeAlias
to use a different name for standard dataset types, or change the
current dataset that is being operated with a single change (e.g.
from raw to calexp).
All storage is still done using the 'real' datasetType name (not
the alias type).