butler: use temporary filename in LogicalLocation
We were writing outputs of type FitsStorage to the originally intended
filename (inside logLoc) instead of the temporary filename, and the
written data was getting clobbered when we move the (empty) temporary
file to the originally intended filename. We need to use the temporary
filename in logLoc. That also simplifies all the other cases (don't
need to use SafeFile), so I've reverted them back to what they were
before we started down this road.

Conflicts:
	python/lsst/daf/persistence/butler.py

This is essentially a cherry-pick of HSC-1321's:
https://github.com/HyperSuprime-Cam/daf_persistence/commit/55ab894375488ed7d49a218aa8a81333226214ce