Registry lookup support.
Adds lookup access through access and posixStorage. Updates
registries.py so that the sqlite3-or-posix logic is in the factory
function constructor.
Adds commentage regarding regestries.py currently working only on
the local filesystem, and notes that it could be updated at least
for the sqlite3 registry, but is out of scope for the current work.