fix mapper instantion for hscMapper
hscMapper's init signature was __init__(self, **kwargs)
which breaks the try/catch TypeError mechansim I was using
to differentiate between old & new style mapper init (to
use cfg or not).
This now uses inspect to see if the mapper's signature has
a 'cfg' kwarg, and if so trys the new style init.