Refactor check for iterable
On Python 3 str has __iter__ so all the tests that are checking
for an iterable in the codebase have to also be careful not to
include str by mistake. Refactor the code that looks for __iter__
to use routines from utils where possible. Where not possible
add instance checks for string.