fix mapperArgs add algorithm and add parent repo for write only repos
mapperArgs adding vars needed to add vars for base classes only
when the subclass has a **kwargs iarg.

added concept of 'unreadable' repos (it's a bad name, because it's
also unwritable - I'm looking for a generic name for this) that
inits the repo so that its registry can be fetched and passed to
child repos.