fix test ButlerSubsetTestCase.testSingleIteration
The new Butler init sequence requires that input repositories
exist. This test was using an input that did not exist, and
it did not matter that the repo did not exist, because the
test is entirely for processing mapper data, that is hard
coded into the test mapper for this test.

Fix is to make the the repo used by the Butler for this test
be an output with mode='rw' instead of an input with mode='r'.