make tests conform to parent rule in Butler
The parent rule is "parents of writable repositories must
be inputs". This might end up being a little tedious, as
if an existing output is loaded a second time, all the
parents of that output must exactly match the inputs
passed to Butler.__init__. This might be acceptable in
our environment as reloaded Butler output repositories
will be loaded by scripts that I think will load all the
same repos as needed multiple times.
If it's not ok, we can add a stage that preloads the
parents of outputs as inputs before the final parents
list is needed to be known.