monkey patch a temp file handle onto downloaded datasets
When a dataset is downloaded from a remote resource, SwiftStorage
puts the dataset into a NamedTemporaryFile, which will be deleted
when the file object is deleted (or closed explicitly). This
guarantees the file will exist as long as the object exists (in
case the object refers internally to the file).