close the database connection in SqlRegistry.__del__
note that for __del__ to get called the garbage collector must
be triggered after the final reference has been released. in
unit tests this may be done in tearDown by calling gc.collect.

If we find having or closing the open database connection is a
problem in production code, we may need to
add api to butler to explicity release database connections (and
maybe other things like in-memory cached objects).